chore: rename to sign-utils
